ai - study - assistant/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
bcrypt>=4.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pymupdf==1.23.8
httpx==0.25.2
python-dotenv==1.0.0
Pillow>=10.0.0
google-auth>=2.0.0
pydantic[email]>=2.0.0
weasyprint>=60.0
